// src/entities/block/model/types.ts
export type BlockType =
  | "text" // 글 - 문장/인용
  | "image" // 이미지
  | "work" // 작품
  | "link" // 링크모음
  | "notice" // 공지/일정
  | "challenge" // 글쓰기챌린지
  | "event" // 이벤트
  | "sns" // SNS
  | "calendar"
  | "blank"
  | "book"; // 여백
// ...
export type TemplateType =
  | "profile"
  | "magazine" /* | "portfolio" | "magazine" | "visual" */;